Heading-based flight, script-driven ship loading, build reach, construction sweeps (play-test round 4)
Aircraft
- Flight now follows the nose: speed builds along the heading, the
  heading turns at the unit's FBI turn rate, and only the final stretch
  to a destination steers directly. Attack runs fly flat out along the
  nose, so approach, run-out and turn-back are banked arcs. The bank
  angle is eased in the sim (UnitPhysicsInfoAir::roll) and the renderer
  adds pitch from the climb rate, replacing the per-tick snap.

Transports
- Sea transports load and unload through their own scripts:
  TransportPickup(unit) and TransportDrop(unit, packedXZ) open the doors,
  swing the crane out, hook the unit (attach-unit), stow it in the hold
  (piece -1, hidden) and set it down again (drop-unit). A 10 s fallback
  covers scripts that never attach or release. Verified on the Hulk.
- COB VM fixes needed for that, which affect every script:
  call-script / start-script passed arguments in reverse order; packed
  x/z positions were unpacked as unsigned, turning negative offsets into
  huge distances; attach-unit carries three operands.

Construction
- Builders measure reach to the target's footprint and path to its
  nearest edge, instead of walking round to the front.
- The green fill sweeps up over the middle third; the texture sweep
  finishes by ~92%.

Also: tools/visual-test.ps1, the PowerShell harness used to drive the
game and capture the construction, flight and Hulk sequences.

+++ src/rwe/cob/cob_util.cpp
@@ -15,8 +15,10 @@ namespace rwe
 
     std::pair<CobPosition, CobPosition> cobUnpackCoords(uint32_t xz)
     {
-        auto x = CobPosition::fromInt(static_cast<int>(xz >> 16));
-        auto z = CobPosition::fromInt(static_cast<int>(xz & 0xffff));
+        // Both halves are signed 16-bit: scripts subtract packed positions to
+        // get an offset, and an offset to the west or north is negative.
+        auto x = CobPosition::fromInt(static_cast<int16_t>(static_cast<uint16_t>(xz >> 16)));
+        auto z = CobPosition::fromInt(static_cast<int16_t>(static_cast<uint16_t>(xz & 0xffff)));
         return std::make_pair(x, z);
     }
